DatagramChannel should not restrict usage to InetSocketAddress
Motivation:

At the moment DatagramChannel and DatagramPacket requires InetSocketAdress which limits the flexibility as for example there could also be a DatagramChannel that uses unix domain sockets.
Beside this there are also still methods present in the public API that take a Promise as argument. These should be removed

Modifications:

- Make DatagramChannel & Datagram packet use SocketAddress
- Remove method overloads that take Promise
- Adjust code for the API changes

Result:

More flexible interface / cleanup